// frontend/scratch/old-src/frontend/src/shared/lib/types.ts
export type ForumCategoryResponse = {
  id: number;
  name: string;
  slug: string;
  description: string;
  threadCount: number;
  postCount: number;
};